implemented requested changes
